Fix 27 (#32)
* fix: #27

Did several things to hopefully fix #27:

 - Switched from \`MathJax.Hub\` to \`MathJax.typeset\`
 - Added a wait for MathJax, rather than just letting jQuery return an asynchronous/deferred loading warning
 - Made sure that Tex On/Off actually work as intended (i.e. when you click \`Tex On\`, you add disabled to \`Tex On\`, not \`Tex Off\`)
 - Added raw HTML disabled props to disabled buttons

* feat(problemsuggestion_form): automatically update MathJax on change

This commit automatically forces MathJax to re-render as a user types, rather than on a defocus, allowing a user to be able to have a live preview.

* feat(problemsuggestion_form): add auto-tex on/off

Since MathJax is known for being slow, we'll allow Auto-Texing to be turned off, just in case we have a very long solution/statement that kills MathJax.

* Reword, lint

Co-authored-by: Evan Chen <evan@evanchen.cc>