Change the AoPS link
Changed the button which leads to the AoPS thread on the problem. In the past it included the entire link (it was shown) and this was ugly and long (leading the button to sometimes even jump off the page). Anyways the user has no use of seeing the link so why no change it to say "Visit AoPS"